FROMGIT: cgroup_freezer: cgroup_freezing: Check if not frozen
__thaw_task() was recently updated to warn if the task being thawed was
part of a freezer cgroup that is still currently freezing:

	void __thaw_task(struct task_struct *p)
	{
	...
		if (WARN_ON_ONCE(freezing(p)))
			goto unlock;

This has exposed a bug in cgroup1 freezing where when CGROUP_FROZEN is
asserted, the CGROUP_FREEZING bits are not also cleared at the same
time. Meaning, when a cgroup is marked FROZEN it continues to be marked
FREEZING as well. This causes the WARNING to trigger, because
cgroup_freezing() thinks the cgroup is still freezing.

There are two ways to fix this:

1. Whenever FROZEN is set, clear FREEZING for the cgroup and all
children cgroups.
2. Update cgroup_freezing() to also verify that FROZEN is not set.

This patch implements option (2), since it's smaller and more
straightforward.

Signed-off-by: Tim Van Patten <timvp@google.com>
Tested-by: Mark Hasemeyer <markhas@chromium.org>
Signed-off-by: Tejun Heo <tj@kernel.org>
(cherry picked from commit 15517a8cb60779e6d0a88983793574f4d7eb866b
 git://git.kernel.org/pub/scm/linux/kernel/git/tj/cgroup.git for-6.7-fixes)

BUG=b:304962611
TEST=powerd_dbus_suspend ; verify no warning in dmesg

Change-Id: I30fa206eb7c03ae6a9087240e9f6c1e547089482
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5044611
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Mark Hasemeyer <markhas@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>