UPSTREAM: soc: mediatek: mtk-svs: Change the thermal sensor device name
This driver tries to create a device link to the thermal sensor device:
change all instances of "lvts" and "thermal" to "thermal-sensor", as
that's what the devicetree node name must be.

Note for MT8183: As specified in a previous commit, this SoC never got
SVS probing, so this is not a breaking change and it does not require
fallback for older device trees.

Link: https://lore.kernel.org/r/20231121125044.78642-7-angelogioacchino.delregno@collabora.com
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 30d83ef88feb79a3c9c940c404b30bd351d9f327)

BUG=b:326193888
TEST=dmesg

Change-Id: I2e67d2b4bec17403b2fdd5f99084f192f0a7ae66
Signed-off-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5364151
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>