UPSTREAM: soc: mediatek: mtk-svs: Drop supplementary svs per-bank pointer
Drop the "pbank" pointer from struct svs_bank: this was used to simply
pass a pointer to the SVS bank that the flow was working on.
That for instance needs more locking, and it's avoidable by adding one
more parameter to functions working on specific banks, either a bank
index number, or passing the svs_bank pointer directly from the caller.

Even if the locking can now be reduced, for now, it was still left in
place for the sake of making sure to not introduce any stability and/or
reliability regression.

Link: https://lore.kernel.org/r/20231121125044.78642-12-angelogioacchino.delregno@collabora.com
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 63077f99b18bedd06f843057ac5b1bbe290e06a1)

BUG=b:326193888
TEST=dmesg

Change-Id: Icc57f8726be6ab33c96824a4aea12f881238faf7
Signed-off-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5365365
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>