FROMGIT: drm/i915/dp: Abort AUX on disconnected native DP ports
An AUX transfer on any disconnected DP port results in long
timeout/retry delays the same way as this is described for TypeC port in

commit a972cd3f0eb5 ("drm/i915/tc: Abort DP AUX transfer on a disconnected TC port")

Prevent the delay on non-TypeC ports as well by aborting the transfer if
the port is disconnected. For eDP keep the current behavior as the
support for HPD signaling is optional for it.

Link: https://patchwork.freedesktop.org/patch/msgid/20240104083008.2715733-13-imre.deak@intel.com
Reviewed-by: Jouni Högander <jouni.hogander@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
(cherry picked from commit d544d000a3cbf845825508dafe0aebe7f93f0e04
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:231655517
TEST=Dp-alt and dp-tunnel display works without any issues

Change-Id: Id20268e565c6b96699fb1b6dcedc25e93477c2cd
Signed-off-by: vivek trambadiya <vivek.dilipbhai.trambadiya@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5253516
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Khaled Almahallawy <khaled.almahallawy@intel.com>
Tested-by: Waheem Md <waheem.md@intel.com>