FROMGIT: lkdtm: Make lkdtm_do_action() return to avoid tail call optimization
The comments for lkdtm_do_action() explicitly call out that it
shouldn't be inlined because we want it to show up in stack
crawls. However, at least with some compilers / options it's still
vanishing due to tail call optimization. Let's add a return value to
the function to make it harder for the compiler to do tail call
optimization here.

Now that we have a return value, we can actually use it in the
callers, which is a minor improvement in the code.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/r/20240122164935.1.I345e485f36babad76370c59659a706723750d950@changeid
Signed-off-by: Kees Cook <keescook@chromium.org>
(cherry picked from commit 189a4edb774b9ee5daf345e703e53fffaf509285
 https://git.kernel.org/pub/scm/linux/kernel/git/kees/linux.git for-next/hardening)

BUG=b:309669058
TEST=crash.KernelCrash.hung_task

Change-Id: I345e485f36babad76370c59659a706723750d950
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5225726
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>