CHROMIUM: iwl7000: mvm: don't always disable eSR due to BT coex
2.4 GHz/LB (low band) link can't be used in an eSR links pair when
BT is on. But eSR is still allowed for a pair of links which none of
them operates in LB.
In the existing code, eSR will always be disabled if one of the
usable links is in LB (and BT is on).
Move this check to the code that verifies a specific pair of links,
and only if one of these links operates on LB - disable eSR.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ib1ba1cd1fa1f026fceec12f89a172fdebe732b96
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: f2ec1c40df010082138537e1a24e4e1b684fcb36