FIXUP: CHROMIUM: kvm: x86: upper bound for boost duration
Commit 1d93a98343c1 ("FIXUP: CHROMIUM: kvm: x86: paravirt sched framework")
introduced a bug which broke the throttling logic.
kvm_arch_vcpu_set_sched_attr is used to record the scheduling
attributes and boost status in arch->pv_sched. It takes the new
scheduling attributes and compare with existing attributes before
setting the new value. A refactoring caused the setting of the
attributes before the comparison happens and this caused the updation
to never happen.

Move the comparison logic back and compare the normalized priority of
previous and new scheduling attributes to check if it needs to be
updated.

UPSTREAM-TASK=b:303645537
BUG=b:262267726
TEST=boot

Fixes: 1d93a98343c1 ("FIXUP: CHROMIUM: kvm: x86: paravirt sched framework")
Change-Id: Ifcb2dac22426b0c61bcd69539f08000ccf28a1f2
Signed-off-by: Vineeth Pillai (Google) <vineeth@bitbyteword.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5308315
Tested-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Joel Fernandes <joelaf@google.com>