CHROMIUM: iwl7000: mvm: flush async_handlers_wiphy_wk on stop
During the stop flow, the async_handlers_wk is flushed but the
async_handlers_wiphy_wk is not, leading to a potential issue where the
scan status may not be cleared, resulting in a warning in
__iwl_mvm_mac_stop. This occurs because the scan_complete_umac
notification handler is now running as a wiphy work and may not be
scheduled before the status check. To fix this, flush the notification
handlers that run as a wiphy work as well.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I6cdb4b1a68c95b7ddf53b791cc980c981754b3d4
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b079e954b45a5f3a22812786127dc1178e713d8e