UPSTREAM: regmap: Add regmap_read_bypassed()
Add a regmap_read_bypassed() to allow reads from the hardware registers
while the regmap is in cache-only mode.

A typical use for this is to keep the cache in cache-only mode until
the hardware has reached a valid state, but one or more status registers
must be polled to determine when this state is reached.

For example, firmware download on the cs35l56 can take several seconds if
there are multiple amps sharing limited bus bandwidth. This is too long
to block in probe() so it is done as a background task. The device must
be soft-reset to reboot the firmware and during this time the registers are
not accessible, so the cache should be in cache-only. But the driver must
poll a register to detect when reboot has completed.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Fixes: 8a731fd37f8b ("ASoC: cs35l56: Move utility functions to shared file")
Link: https://msgid.link/r/20240408101803.43183-2-rf@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 70ee853eec5693fefd8348a2b049d9cb83362e58)

BUG=b:296322074
TEST=Test Audio use cases.

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Idd7dfb94529b777642b539a6f632d68fc0cf6fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5514578
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>