CHROMIUM: media: mediatek: decoder: Limit to 1080p for normal playback
There's a known issue that system gets janky and hang due to heavy
memory demand for 2K and 4K video decoding through MTK HW vcodec.

However, the 4K decoding capability is needed to enable the SVP feature.

This is a workaround to limit the normal world video decoding to 1080p.
Eventually we want to revert this once the culprit is found and fixed.

BUG=b:332235023
TEST=system doesn't hang when browsing a folder full with 4K videos in
     Files app
UPSTREAM-TASK=b:339132693

Change-Id: I2d10a2467a02929f718a9010aca63e7bfaea60db
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5513912
Reviewed-by: Pin-yen Lin <treapking@chromium.org>