CHROMIUM: arm64: dts: mt8188: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading
the power, the display power is calculated in watts due to
unsynchronized reading and writing. Use virtual battery 0xf
to communicate with EC can solve this.

BUG=b:339132173
TEST=check dispaly battery

Change-Id: Ib294061696acc476468cc226b06e94802df9ba5d
Signed-off-by: Cong Yang <yangcong5@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5524614
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Rui Zhou <zhourui@huaqin.corp-partner.google.com>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>