CHROMIUM: thermal: mediatek: Dup thermal driver for MT8192
MT8192 is being uprev'd to v6.1 and carry ancient downstream thermal
driver from v5.4 branch, which conflicts with the newer version that
will be used for MT8188.

To make sure this won't block MT8188 development, duplicate and rename
the driver to ensure 6.1 can have two parallel thermal solution for two
different boards.

This is definitely a technical debt that we don't want to have in the
first place, but we don't have much time to fix it at this moment, and
eventually this will be fully replaced by the upstream solution when
that becomes available, hopefully within a few months.

BUG=b:236331727
TEST=emerge-geralt chromeos-kernel-6_1
UPSTREAM-TASK=b:326555352

Change-Id: If8238d81b3eaac1ae56aff17974f122eec4bec93
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5526391
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Tested-by: Hsin-Te Yuan <yuanhsinte@chromium.org>