FIXUP: FROMLIST: drm/mediatek: Set DRM mode configs accordingly
As mentioned in [1], we don't need a full-size cursor plane, which ended
up letting userspace allocates two 64MB buffers instead of two 1MB ones
and cause additional memory pressure on Geralt.

Set the cursor plane size down to 512. Note that the DRM framework
uses 64 as default, but IGT may test the cursor size up to 512x512.

The series is not fully upstreamed so I expect this will be fixed in the
next FROMLIST revision.

BUG=b:342940839
TEST=grep dma-coherent /proc/vmallocinfo | \
         awk '{total+=$2}; END {print total/1024/1024}'
     the allocated buffer size go down
TEST=Two DMA buffer allocation in mtk_gem_create() go down from 64MB to 1MB
UPSTREAM-TASK=b:287363873

[1]: https://lore.kernel.org/all/CAC=S1ngUh-o107YM8zEu-ebaP7+eaO1x6SsZ6OeRxrH5YgFKxw@mail.gmail.com/

Change-Id: I6740fd85b8b2f568d54b6e934d711f9233701283
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5568704
Reviewed-by: Wei-Shun Chang <weishunc@chromium.org>